requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-ai>=0.0.1

# AWS
boto3>=1.28.0
botocore>=1.31.0

# AI/LLM
openai>=1.0.0

# Anthropic Claude (optional)
anthropic>=0.18.0

# Video RAG system
chromadb>=0.4.0
assemblyai>=0.17.0

# Utilities
python-dotenv>=1.0.0
